Change the config structure, fix GUI support
So! Apparently I misunderstood the purpose of `config.schema.json`.
Let's fix that.

But, apparently there's no way to represent an object with arbitrary keys
in Angular JSON Schema Form markup (or I can't figure it out, anyway).
This means I can't make a working form using the existing schema.

As such, version 2.0! This is a breaking change.