Fix already-known bridge check; optionally access properties in responses
When I moved to the BridgeManager design, I deprecated the plugin-owned
map of bridges, but I didn't remove it or the stale check into it. This
adds a helper method to the BridgeManager for the purposes of seeing if
a newly-announced bridge is already known.

Also, the SerenaTiltOnlyWoodBlind class' handler for unsolicited
responses wasn't defensive enough, and assumed everything was for it
(and shaped correctly). This change checks for properties, and will
silently discard unsolicited messages if they're not shaped right.