Explicitly close the bridge if we already have one.
Bridges spring forth fully formed from the foreheads of the
BridgeFinder, with a connection and a ping loop and everything. That
aforementioned ping loop inconveniently keeps the bridge (and its live,
active) connection around even if we don't store it in the bridge
manager.

This change expicitly closes the bridge, which should permit it to
finally go away.