Check for lower-case bridge IDs when one is announced
When checking to see if an announced bridge is already known, convert
its ID to lower case before looking in the map. When a SmartBridge is
created, its ID is lower-cased, and that lower-cased value is then
stored as the map key. When a new bridge comes along, we don't
intentionally lower-case the ID when looking it up, which causes a miss.

This is an annoying bug. Before the pairing UI, bridge IDs had to be
lower case to match the ID extracted from the hostname. Now we ask the
bridge for its ID directly, and get it in upper case. But people have
existing configs that use lower case. If I were smart, I'd refactor it
so BridgeInfo passes lower case, but that's going to take time to get it
right. So anyway, that's the story.