Support filtering Picos, disabling blinds, and long- and double-press
Okay, so the big one is that this adds support for long- and
double-press events for Pico remotes. The timing is configurable in an
effort to make the feature maximally accessible.

You can also turn off Serena Tilt-Only Wood Blinds support in the
plugin. This is useful now that they're natively supported. This is not,
however, the default behavior, because that could break existing
installs.

Lastly, you can now filter out (hide from HomeKit) Pico remotes that are
set up in the Lutron app or paired to devices. Some people don't want to
be able to use them in both places, and that's groovy.