Fix filtered devices that don't disappear
This change refactors the startup code so that filtered devices are
removed from HomeKit. This is done by storing cached accessories, and
then unregistering them when they're detected on the bridge. Cached
accessories that are *not* discovered on the bridge won't be removed,
however. Cache invalidation is hard.

Anyway, this should also make startup a lot less brittle, as
constructors are now synchronous and initialization is async, and we
treat future things in the future.

Fixes #10
Fixes #65