Refactor to reduce scary log lines and noise
This refactor changes how accessories are initialized. Now,
initialization returns an indicator of the status of wiring it up:
either success, skipped, or error. Handling that outcome is now done
in-place. This is the latest in the series I call "Rust-flavored
TypeScript". Anyway, now we only show error lines when there are errors.