Make the plugin re-subscribe on disconnection
When the client connection is lost, the bridge removes subscriptions to
events, which makes sense. The plugin only subscribes to events when it
wires up the accessories, and never again.

This change listens for the disconnected event, and makes the plugin
re-subscribe when that happens. This should keep devices requiring
subscription, namely Picos and Occupancy Sensors, working after client
connection is re-established.