Replace connections to announced bridges
This is yet another attempt to resolve #31.

This change alters behavior when a bridge we've already seen
re-announces itself on the network. This can happen periodically either
because, perhaps, the bridge restarted itself, updated its firmware, was
diconnected and reconnected from the network, or just because it felt
like it. Previously, we'd ignore the announcement, and rely on the
LeapClient to reconnect if needed, but this obviously did not work. So
now we create an entirely new connection, and stick it into the bridge.

An initial implementation tried to replace the whole bridge object, and
add an event to the LutronCasetaLeap platform device to propagate the
new bridge object, but that didn't really work mostly because the
OccupancySensorRouter has some messy-ish handler code that would likely
keep references to the old bridge around (and would require
re-initialization).